raw conversion of FakeConfigurator specs
