+++ Source/FakeItEasy.Specs/FakeItEasy.Specs.csproj
@@ -89,7 +89,7 @@
     <Compile Include="GlobalSuppressions.cs" />
     <Compile Include="Creation.cs" />
     <Compile Include="DomainEvents.cs" />
-    <Compile Include="FakeConfiguratorSpecs.cs" />
+    <Compile Include="FakeConfigurator.cs" />
     <Compile Include="SelfInitializedFakesSpecs.cs" />
     <Compile Include="Lazy.cs" />
     <Compile Include="DummyFactory.cs" />
